+++ src/macos/mod.rs
@@ -1,4 +1,3 @@
 mod macos_impl;
 
-pub mod keycodes;
 pub use self::macos_impl::Enigo;